+++ libs/ui/src/components/atoms/Badge/Badge.types.ts
@@ -1,6 +1,6 @@
 import React, { ReactNode } from 'react';
 
-export type BadgeColor = 'blue' | 'green' | 'pink' | 'violet';
+export type BadgeColor = 'blue' | 'green' | 'pink' | 'violet'; // TODO Create shared types file in root of atoms. Share between this and Tag
 export type BadgeSize = 'sm' | 'lg';
 
 export interface BadgeProps {
