+++ libs/ui/src/components/atoms/Label/Label.tsx
@@ -4,6 +4,11 @@ import { StyledLabel } from './Label.styles';
 export const Label: React.FC<LabelProps> = ({
   children = 'label',
   id,
+  className,
 }: LabelProps) => {
-  return <StyledLabel htmlFor={id}>{children}</StyledLabel>;
+  return (
+    <StyledLabel htmlFor={id} className={className}>
+      {children}
+    </StyledLabel>
+  );
 };
