+++ libs/ui/src/components/atoms/Label/Label.types.ts
@@ -1,4 +1,5 @@
 export type LabelProps = {
   id?: string;
   children: React.ReactChild;
+  className?: string;
 };
