+++ libs/ui/src/components/atoms/Link/Link.types.ts
@@ -1,0 +1,17 @@
+import { IconType } from 'react-icons';
+
+export interface LinkProps extends React.ComponentPropsWithRef<'a'> {
+  href?: string;
+  Icon?: IconType;
+  LeftIcon?:
+    | IconType
+    | React.FunctionComponent<
+        React.SVGProps<SVGSVGElement> & {
+          title?: string | undefined;
+        }
+      >;
+  selected?: boolean;
+  disabled?: boolean;
+  linkType?: 'internal' | 'external' | 'no-icon-external';
+  hideIcon?: boolean;
+}
