+++ libs/ui/src/components/atoms/Link/index.ts
@@ -1,1 +1,5 @@
-export * from './Link';
+import { Link } from './Link';
+import type { LinkProps } from './Link.types';
+
+export type { LinkProps };
+export { Link };
