+++ libs/ui/src/components/atoms/Loading/Loading.types.ts
@@ -1,0 +1,6 @@
+import { ButtonColor } from '../Button/Button.types';
+
+export type LoadingProps = {
+  size?: number;
+  color?: ButtonColor;
+};
