+++ libs/ui/src/components/atoms/Loading/index.ts
@@ -1,1 +1,5 @@
-export * from './Loading';
+import { Loading } from './Loading';
+import type { LoadingProps } from './Loading.types';
+
+export type { LoadingProps };
+export { Loading };
