+++ libs/ui/src/components/atoms/Progress/Progress.types.ts
@@ -1,0 +1,7 @@
+export type ProgressProps = {
+  backgroundColor: string;
+  progressSection: {
+    percentage: string;
+    color: string;
+  }[];
+};
