+++ libs/ui/src/components/atoms/Progress/index.ts
@@ -1,1 +1,5 @@
-export * from './Progress';
+import { Progress } from './Progress';
+import type { ProgressProps } from './Progress.types';
+
+export type { ProgressProps };
+export { Progress };
