+++ libs/ui/src/components/atoms/Radio/Radio.tsx
@@ -1,9 +1,7 @@
 import React from 'react';
-import {
-  RadioGroupProps,
-  RadioGroupItemProps,
-} from '@radix-ui/react-radio-group';
 
+import { Label } from '../Label/Label';
+import { RadioItemGroupProps, RadioItemProps } from './Radio.types';
 import {
   RadioGroup,
   RadioItem,
@@ -11,27 +9,13 @@ import {
   Container,
   LabelContainer,
 } from './Radio.styles';
-import { Label } from '../Label/Label';
 
-export type RadioProps = { label: string; id: string } & RadioGroupItemProps;
-export type Props = {
-  className?: string;
-  defaultValue?: string;
-  radios: RadioProps[];
-};
-
-type Ref =
-  | React.RefObject<HTMLInputElement>
-  | ((instance: HTMLInputElement | null) => void)
-  | null
-  | undefined;
-
-export const Radio = React.forwardRef(
-  (props: RadioGroupProps & Props, ref: Ref) => {
+export const Radio = React.forwardRef<HTMLInputElement, RadioItemGroupProps>(
+  (props, ref) => {
     const { className, radios, defaultValue } = props;
     return (
       <RadioGroup {...props} defaultValue={defaultValue} className={className}>
-        {radios.map((radio: RadioProps) => (
+        {radios.map((radio: RadioItemProps) => (
           <Container key={radio.id}>
             <RadioItem {...radio}>
               <RadioIndicator ref={ref} />
