+++ libs/ui/src/components/atoms/Radio/index.ts
@@ -1,1 +1,5 @@
-export * from './Radio';
+import { Radio } from './Radio';
+import type { RadioItemGroupProps } from './Radio.types';
+
+export type { RadioItemGroupProps };
+export { Radio };
