+++ libs/ui/src/components/atoms/Select/Select.types.ts
@@ -1,0 +1,3 @@
+import type { SelectProps } from '../../../types/formAndField';
+// TODO Look into Select Props and move into Select Folder
+export type { SelectProps };
