+++ libs/ui/src/components/atoms/Select/index.ts
@@ -1,1 +1,5 @@
-export * from './Select';
+import { Select } from './Select';
+import type { SelectProps } from './Select.types';
+
+export type { SelectProps };
+export { Select };
