+++ libs/ui/src/components/atoms/Spinner/Spinner.types.ts
@@ -1,0 +1,9 @@
+export type SpinnerType = {
+  topColor?: string;
+  bottomColor?: string;
+  size?: string;
+  strokeWidth?: string;
+  margin?: string;
+  padding?: string;
+  speed?: string;
+};
