+++ libs/ui/src/components/atoms/Spinner/index.ts
@@ -1,1 +1,5 @@
-export * from './Spinner';
+import { Spinner } from './Spinner';
+import type { SpinnerType } from './Spinner.types';
+
+export type { SpinnerType };
+export { Spinner };
