+++ libs/ui/src/components/atoms/Switch/index.ts
@@ -1,1 +1,5 @@
-export * from './Switch';
+import { Switch } from './Switch';
+import type { SwitchProps } from './Switch.types';
+
+export type { SwitchProps };
+export { Switch };
