+++ libs/ui/src/components/atoms/Tag/index.ts
@@ -1,1 +1,5 @@
-export * from './Tag';
+import { Tag } from './Tag';
+import type { TagProps } from './Tag.types';
+
+export type { TagProps };
+export { Tag };
