+++ libs/ui/src/components/atoms/TextArea/TextArea.types.ts
@@ -1,0 +1,6 @@
+import { Field } from '../../../types/formAndField';
+
+export type TextAreaProps = Field & {
+  className?: string;
+  height?: string;
+};
