+++ libs/ui/src/components/atoms/TextArea/index.ts
@@ -1,1 +1,5 @@
-export * from './TextArea';
+import { TextArea } from './TextArea';
+import type { TextAreaProps } from './TextArea.types';
+
+export type { TextAreaProps };
+export { TextArea };
