+++ libs/ui/src/components/atoms/Tooltip/Tooltip.types.ts
@@ -1,0 +1,8 @@
+export type TooltipProps = {
+  content?: string | React.ReactNode;
+  side?: 'top' | 'right' | 'bottom' | 'left';
+  triggerEl?: React.ReactNode;
+  offset?: number;
+  delay?: number;
+  triggerAsChild?: boolean;
+};
