+++ libs/ui/src/components/atoms/Tooltip/index.ts
@@ -1,2 +1,6 @@
-export * from './Tooltip';
-export * from './TooltipTriggers';
+import { Tooltip } from './Tooltip';
+import { SmTooltipIcon, LgTooltipIcon } from './TooltipTriggers';
+import type { TooltipProps } from './Tooltip.types';
+
+export type { TooltipProps };
+export { Tooltip, SmTooltipIcon, LgTooltipIcon };
